changed location of endModal
