+++ components/Game/Player1HP.js
@@ -34,49 +34,6 @@ const Player1HP = (props) => {
 	});
 
 	useEffect(() => {
-		if (ended) {
-			console.log('THE GAME HAS ENDED FROM USE EFFECT');
-			const setEndState = () => {
-				if (health1 > health2) {
-					dispatch(gameActions.setWinner(player1.id));
-				} else if (health2 > health1) {
-					dispatch(gameActions.setWinner(player2.id));
-				}
-				setEndState();
-				console.log('game winner', winner);
-				props.setEndModal(true);
-				const updateGameHistory = async () => {
-					if (health1 === health2) {
-						try {
-							await supabase.from('game').update({ isDraw: true }).eq('id', Router.query.id);
-						} catch (error) {
-							console.log('error in update game history', error);
-						}
-					} else {
-						try {
-							await supabase.from('game').update({ winner: winner }).eq('id', Router.query.id);
-						} catch (error) {
-							console.log('error in update game history', error);
-						}
-						const { data } = await supabase.from('users').select('wallet').eq('id', winner);
-						console.log('winner wallet value', data);
-						const walletUpdate = data + 3;
-						const updateWallet = async () => {
-							try {
-								await supabase.from('users').update({ wallet: walletUpdate }).eq('id', winner);
-							} catch (error) {
-								console.log('error updating winner wallet', error);
-							}
-						};
-						updateWallet();
-					}
-				};
-				updateGameHistory();
-			};
-		}
-	}, [ended]);
-
-	useEffect(() => {
 		health1 <= 0 && dispatch(gameActions.endGame(true));
 		health1 <= 0 && dispatch(loadActions.setLoading(true));
 	}, [health1]);
